feat: add comprehensive MCP test suite and professional Makefile
- Add complete integration test covering all MCP tool functions
  * TestMCPToolFunctions: node/edge operations, queries, deletions, error handling
  * TestMCPCommandValidation: parameter validation and edge cases
  * Complex scenario testing with multi-node project graphs
- Create professional Makefile with 30+ development targets
  * Essential: help, build, test, test-mcp, validate
  * Testing: unit, integration, benchmark, coverage, race detection
  * Workflow: lint, fmt, vet, clean, pre-commit
  * Execution: run, run-persistent, dev, mcp-demo
- Add MCP command constants to eliminate magic strings
- Fix linter issues: parameter ordering, unnecessary conversions
- Update CLAUDE.md with comprehensive Makefile usage guide
- Pre-allocate slices for better performance in query operations

All tests pass with comprehensive MCP protocol coverage.

+++ internal/storage/bolt.go
@@ -259,7 +259,7 @@ func (b *BoltBackend) updateStats() {
 		// Get database file size (simplified approach)
 		if stat := b.db.Stats(); stat.TxStats.PageCount > 0 {
 			// Use a reasonable page size estimate
-			b.stats.DatabaseSize = int64(stat.TxStats.PageCount) * 4096
+			b.stats.DatabaseSize = stat.TxStats.PageCount * 4096
 		}
 
 		// Count nodes
